Revert "Automatically register VMInstalls for the JDKs installed on the local machine (#3251)"
This reverts commit 6cc071a5d795780907109ea36fa2b9f0e25bebf2.
