+++ package.json
@@ -1529,7 +1529,7 @@
     "fs-extra": "^8.1.0",
     "glob": "^7.1.3",
     "htmlparser2": "6.0.1",
-    "jdk-utils": "^0.5.0",
+    "jdk-utils": "^0.4.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "semver": "^7.5.2",
